tadabbur-data/requirements.txt
fastapi==0.115.0
uvicorn==0.30.0
neo4j==5.14.1
psycopg2-binary==2.9.9
beautifulsoup4==4.12.2
requests==2.31.0
python-dotenv==1.0.0
pydantic==2.7.0
python-multipart==0.0.9
google-generativeai